finally made the chuncks works , need to be tested

+++ server/Server.cpp
@@ -61,7 +61,7 @@ std::string Server::getRecvBuffer(){
 	// return res;
 }
 
-size_t Server::getRecvBufferLenght(){
+int Server::getRecvBufferLenght(){
 	return this->recvBuffer.str().size();
 }
